+++ engine/src/hopeit/cli/server.py
@@ -2,10 +2,11 @@
 CLI server commands
 """
 import sys
+from typing import List
 
 try:
     import click
-    from hopeit.server import web
+    from hopeit.server import wsgi
 except ModuleNotFoundError:
     print("ERROR: Missing dependencies."
           "\n       To use hopeit_server command line tool"
@@ -19,28 +20,38 @@ def server():
 
 
 @server.command()
-@click.option('--config-files', prompt='Config files',
+@click.option('--config-files', required=True,
               help='Comma-separated config file paths, starting with server config, then plugins, then apps.')
-@click.option('--api-file', help='Path to openapi complaint json specification.')
+@click.option('--api-file', default=None, help='Path to openapi complaint json specification.')
 @click.option('--host', default='0.0.0.0', help='Server host address or name.')
-@click.option('--port', default='8020', help='TCP/IP port to listen.')
+@click.option('--port', default=8020, help='TCP/IP port to listen.')
 @click.option('--path', help='POSIX complaint socket name.')
 @click.option('--start-streams', is_flag=True, default=False, help='Auto start reading stream events.')
 @click.option('--enabled-groups', default='',
               help="Optional comma-separated group labels to start. If no group is specified, all events will be"
               " started. Events with no group or 'DEFAULT' group label will always be started. 'DEFAULT' group label"
               " can also be used explicitly to start only events with no group or 'DEFAULT' group label.")
-def run(config_files: str, api_file: str, host: str, port: int, path: str, start_streams: bool, enabled_groups: str):
+@click.option('--workers', default=1, help="Number of workeres to start. Max number of workers is (cpu_count * 2) + 1")
+@click.option('--worker-class', type=click.Choice(['GunicornWebWorker', 'GunicornUVLoopWebWorker']),
+              default="GunicornWebWorker", help="Gunicorn aiohttp worker class. Default value is GunicornWebWorker.")
+def run(config_files: str, api_file: str, host: str, port: int, path: str,
+        start_streams: bool, enabled_groups: str, workers: int, worker_class: str):
     """
     Runs web server hosting apps specified in config files.
     """
-    web.prepare_engine(
-        config_files=config_files.split(','),
+    groups: List[str] = [] if enabled_groups == "" else enabled_groups.split(',')
+    files: List[str] = config_files.split(',')
+
+    wsgi.run_app(
+        host=host,
+        port=port,
+        path=path,
+        config_files=files,
         api_file=api_file,
-        enabled_groups=enabled_groups.split(',') if enabled_groups else [],
         start_streams=start_streams,
-    )
-    web.serve(host=host, path=path, port=port)
+        enabled_groups=groups,
+        workers=workers,
+        worker_class=worker_class)
 
 
 cli = click.CommandCollection(sources=[server])